src/requirements.txt
requests
lxml
psycopg2-binary
fastapi
uvicorn[standard]
openpyxl
pandas
python-dotenv
